backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
librosa==0.10.1
pydub==0.25.1
openai-whisper==20231117
ollama==0.1.6
numpy==1.26.3
scipy==1.12.0
python-dotenv==1.0.0
